feat: refactor mapper
