+++ ProductService/ProductService/Mapper/IMapper.cs
@@ -1,0 +1,7 @@
+﻿namespace ProductService.Mapper
+{
+    public interface IMapper
+    {
+        public TOut Map<TIn, TOut>(TIn valueToConvert);
+    }
+}
