+++ ProductService/ProductService/Mapper/Mapper.cs
@@ -1,23 +1,24 @@
 ﻿using AutoMapper;
 using MoneyTypes;
 
-namespace ProductService
+namespace ProductService.Mapper
 {
-    public static class Mapper
+    public class Mapper : ProductService.Mapper.IMapper
     {
-        public static TOut Map<TIn, TOut>(TIn valueToConvert)
+        public TOut Map<TIn, TOut>(TIn valueToConvert)
         {
-            var config = new MapperConfiguration(cfg => {
+            var config = new MapperConfiguration(cfg =>
+            {
                 cfg.CreateMap<TIn, TOut>();
-                // настройки маппинга кастомного типа
+                // configs mapper for custom types
                 cfg.CreateMap<DecimalValue, decimal>().ConvertUsing(val => val);
-                cfg.CreateMap<decimal,DecimalValue>().ConvertUsing(val => val);
+                cfg.CreateMap<decimal, DecimalValue>().ConvertUsing(val => val);
             });
             var mapper = new AutoMapper.Mapper(config);
 
             var result = mapper.Map<TOut>(valueToConvert);
             return result;
-            
+
 
         }
     }
