+++ ProductService/ProductService/Services/ProductServiceGrpc.cs
@@ -1,5 +1,6 @@
 ﻿using Grpc.Core;
 using Product;
+using ProductService.Mapper;
 
 namespace ProductService.Services
 {
@@ -9,10 +10,12 @@ namespace ProductService.Services
     public class ProductServiceGrpc : Product.ProductService.ProductServiceBase
     {
         private readonly ProductContext _context;
+        private readonly IMapper _mapper;
 
-        public ProductServiceGrpc( ProductContext context)
+        public ProductServiceGrpc(ProductContext context, IMapper mapper)
         {   
             _context = context;
+            _mapper = mapper;
         }
 
 
@@ -28,7 +31,7 @@ namespace ProductService.Services
             var products = await  _context.GetAsync();
             products.ForEach(p =>
             {
-                var productResponce = Mapper.Map<Models.Product, GetProductsResponse.Types.Product>(p);
+                var productResponce = _mapper.Map<Models.Product, GetProductsResponse.Types.Product>(p);
                 responce.Products.Add(productResponce);
             }); 
             return responce;
